refactor: slackmessage, ailog 애그리거트 패키지 분리
