+++ notification-service/src/main/java/com/firstlogistics/notificationservice/ailog/domain/entity/AILog.java
@@ -1,12 +1,10 @@
-package com.firstlogistics.notificationservice.domain.entity;
+package com.firstlogistics.notificationservice.ailog.domain.entity;
 
-
-import com.firstlogistics.notificationservice.domain.enums.AILogStatus;
-import com.firstlogistics.notificationservice.domain.enums.MessengerType;
-import com.firstlogistics.notificationservice.domain.exception.AILogErrorCode;
-import com.firstlogistics.notificationservice.domain.exception.AILogException;
-import com.firstlogistics.notificationservice.domain.vo.AILogId;
-import jakarta.persistence.Table;
+import com.firstlogistics.notificationservice.ailog.domain.enums.AILogStatus;
+import com.firstlogistics.notificationservice.ailog.domain.enums.MessengerType;
+import com.firstlogistics.notificationservice.ailog.domain.exception.AILogErrorCode;
+import com.firstlogistics.notificationservice.ailog.domain.exception.AILogException;
+import com.firstlogistics.notificationservice.ailog.domain.vo.AILogId;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -16,13 +14,13 @@ import java.util.UUID;
 @Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public class AILog {
-    AILogId id;
-    UUID messageId;
-    String requestContent;
-    String responseContent;
-    String systemPrompt;
-    AILogStatus status;
-    MessengerType messengerType;
+    private final AILogId id;
+    private UUID messageId;
+    private String requestContent;
+    private String responseContent;
+    private String systemPrompt;
+    private AILogStatus status;
+    private MessengerType messengerType;
 
     public static AILog create(
             String requestContent,
@@ -40,6 +38,18 @@ public class AILog {
                 status,
                 messengerType
         );
+    }
+
+    public static AILog reconstitute(
+            AILogId id,
+            UUID messageId,
+            String requestContent,
+            String responseContent,
+            String systemPrompt,
+            AILogStatus status,
+            MessengerType messengerType
+    ) {
+        return new AILog(id, messageId, requestContent, responseContent, systemPrompt, status, messengerType);
     }
 
     public void updateSlackMessageId(UUID slackMessageId) {
