+++ notification-service/src/main/java/com/firstlogistics/notificationservice/ailog/domain/exception/AILogException.java
@@ -1,4 +1,4 @@
-package com.firstlogistics.notificationservice.domain.exception;
+package com.firstlogistics.notificationservice.ailog.domain.exception;
 
 import common.exception.BaseException;
