--- notification-service/src/main/java/com/firstlogistics/notificationservice/domain/enums/MessengerType.java
@@ -1,5 +1,0 @@
-package com.firstlogistics.notificationservice.domain.enums;
-
-public enum MessengerType {
-    SLACK
-}
