--- notification-service/src/main/java/com/firstlogistics/notificationservice/domain/enums/SlackMessageType.java
@@ -1,6 +1,0 @@
-package com.firstlogistics.notificationservice.domain.enums;
-
-public enum SlackMessageType {
-    DEADLINE,
-    DELIVERY
-}
