+++ notification-service/src/main/java/com/firstlogistics/notificationservice/slackmessage/domain/enums/SlackMessageType.java
@@ -1,0 +1,6 @@
+package com.firstlogistics.notificationservice.slackmessage.domain.enums;
+
+public enum SlackMessageType {
+    DEADLINE,
+    DELIVERY
+}
