+++ notification-service/src/main/java/com/firstlogistics/notificationservice/slackmessage/domain/vo/SlackMessageId.java
@@ -1,4 +1,4 @@
-package com.firstlogistics.notificationservice.domain.vo;
+package com.firstlogistics.notificationservice.slackmessage.domain.vo;
 
 import java.util.UUID;
